third_party/WebKit/LayoutTests/fast/multicol/client-rect-nested.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<div style="position:fixed; z-index:-1; top:8px; left:8px; columns:2; column-fill:auto; height:40px; column-gap:10px; width:630px; line-height:20px; orphans:1; widows:1;">
    <div class="test"><br></div>
    <div style="columns:2; column-gap:10px;">
        <div class="test"><br></div>
        <div style="columns:2; column-gap:10px;">
            <div class="test"><br></div>
            <div class="test"><br></div>
            <div class="test"><br></div>
            <div class="test"><br></div>
            <div class="test"><br></div>
            <div class="test"><br></div>
            <div class="test"><br></div>
            <div class="test"><br></div>
            <div class="test"><br></div>
            <div class="test"><br></div>
        </div>
    </div>
</div>
<script>
    description("Tests getBoundingClientRect() in nested multicol.");
    var elements = document.getElementsByClassName("test");

    shouldBe("elements[0].getBoundingClientRect().left", "8");
    shouldBe("elements[0].getBoundingClientRect().top", "8");
    shouldBe("elements[0].getBoundingClientRect().right", "318");
    shouldBe("elements[0].getBoundingClientRect().bottom", "28");

    shouldBe("elements[1].getBoundingClientRect().left", "8");
    shouldBe("elements[1].getBoundingClientRect().top", "28");
    shouldBe("elements[1].getBoundingClientRect().right", "158");
    shouldBe("elements[1].getBoundingClientRect().bottom", "48");

    shouldBe("elements[2].getBoundingClientRect().left", "168");
    shouldBe("elements[2].getBoundingClientRect().top", "28");
    shouldBe("elements[2].getBoundingClientRect().right", "238");
    shouldBe("elements[2].getBoundingClientRect().bottom", "48");

    shouldBe("elements[3].getBoundingClientRect().left", "248");
    shouldBe("elements[3].getBoundingClientRect().top", "28");
    shouldBe("elements[3].getBoundingClientRect().right", "318");
    shouldBe("elements[3].getBoundingClientRect().bottom", "48");

    shouldBe("elements[4].getBoundingClientRect().left", "328");
    shouldBe("elements[4].getBoundingClientRect().top", "8");
    shouldBe("elements[4].getBoundingClientRect().right", "398");
    shouldBe("elements[4].getBoundingClientRect().bottom", "28");

    shouldBe("elements[5].getBoundingClientRect().left", "328");
    shouldBe("elements[5].getBoundingClientRect().top", "28");
    shouldBe("elements[5].getBoundingClientRect().right", "398");
    shouldBe("elements[5].getBoundingClientRect().bottom", "48");

    shouldBe("elements[6].getBoundingClientRect().left", "408");
    shouldBe("elements[6].getBoundingClientRect().top", "8");
    shouldBe("elements[6].getBoundingClientRect().right", "478");
    shouldBe("elements[6].getBoundingClientRect().bottom", "28");

    shouldBe("elements[7].getBoundingClientRect().left", "408");
    shouldBe("elements[7].getBoundingClientRect().top", "28");
    shouldBe("elements[7].getBoundingClientRect().right", "478");
    shouldBe("elements[7].getBoundingClientRect().bottom", "48");

    shouldBe("elements[8].getBoundingClientRect().left", "488");
    shouldBe("elements[8].getBoundingClientRect().top", "8");
    shouldBe("elements[8].getBoundingClientRect().right", "558");
    shouldBe("elements[8].getBoundingClientRect().bottom", "28");

    shouldBe("elements[9].getBoundingClientRect().left", "488");
    shouldBe("elements[9].getBoundingClientRect().top", "28");
    shouldBe("elements[9].getBoundingClientRect().right", "558");
    shouldBe("elements[9].getBoundingClientRect().bottom", "48");

    shouldBe("elements[10].getBoundingClientRect().left", "568");
    shouldBe("elements[10].getBoundingClientRect().top", "8");
    shouldBe("elements[10].getBoundingClientRect().right", "638");
    shouldBe("elements[10].getBoundingClientRect().bottom", "28");

    shouldBe("elements[11].getBoundingClientRect().left", "568");
    shouldBe("elements[11].getBoundingClientRect().top", "28");
    shouldBe("elements[11].getBoundingClientRect().right", "638");
    shouldBe("elements[11].getBoundingClientRect().bottom", "48");
</script>
