third_party/blink/web_tests/platform/generic/fast/multicol/client-rect-nested-expected.txt

Tests getBoundingClientRect() in nested multicol.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS elements[0].getBoundingClientRect().left is 8
PASS elements[0].getBoundingClientRect().top is 8
PASS elements[0].getBoundingClientRect().right is 318
PASS elements[0].getBoundingClientRect().bottom is 28
PASS elements[1].getBoundingClientRect().left is 8
PASS elements[1].getBoundingClientRect().top is 28
PASS elements[1].getBoundingClientRect().right is 158
PASS elements[1].getBoundingClientRect().bottom is 48
PASS elements[2].getBoundingClientRect().left is 168
PASS elements[2].getBoundingClientRect().top is 28
PASS elements[2].getBoundingClientRect().right is 238
PASS elements[2].getBoundingClientRect().bottom is 48
PASS elements[3].getBoundingClientRect().left is 248
PASS elements[3].getBoundingClientRect().top is 28
PASS elements[3].getBoundingClientRect().right is 318
PASS elements[3].getBoundingClientRect().bottom is 48
PASS elements[4].getBoundingClientRect().left is 328
PASS elements[4].getBoundingClientRect().top is 8
PASS elements[4].getBoundingClientRect().right is 398
PASS elements[4].getBoundingClientRect().bottom is 28
PASS elements[5].getBoundingClientRect().left is 328
PASS elements[5].getBoundingClientRect().top is 28
PASS elements[5].getBoundingClientRect().right is 398
PASS elements[5].getBoundingClientRect().bottom is 48
PASS elements[6].getBoundingClientRect().left is 408
PASS elements[6].getBoundingClientRect().top is 8
PASS elements[6].getBoundingClientRect().right is 478
PASS elements[6].getBoundingClientRect().bottom is 28
PASS elements[7].getBoundingClientRect().left is 408
PASS elements[7].getBoundingClientRect().top is 28
PASS elements[7].getBoundingClientRect().right is 478
PASS elements[7].getBoundingClientRect().bottom is 48
PASS elements[8].getBoundingClientRect().left is 488
PASS elements[8].getBoundingClientRect().top is 8
PASS elements[8].getBoundingClientRect().right is 558
PASS elements[8].getBoundingClientRect().bottom is 28
PASS elements[9].getBoundingClientRect().left is 488
PASS elements[9].getBoundingClientRect().top is 28
PASS elements[9].getBoundingClientRect().right is 558
PASS elements[9].getBoundingClientRect().bottom is 48
PASS elements[10].getBoundingClientRect().left is 568
PASS elements[10].getBoundingClientRect().top is 8
PASS elements[10].getBoundingClientRect().right is 638
PASS elements[10].getBoundingClientRect().bottom is 28
PASS elements[11].getBoundingClientRect().left is 568
PASS elements[11].getBoundingClientRect().top is 28
PASS elements[11].getBoundingClientRect().right is 638
PASS elements[11].getBoundingClientRect().bottom is 48
PASS successfullyParsed is true

TEST COMPLETE
